split ui crates and update clippy suggestions
