+++ Cargo.toml
@@ -12,7 +12,7 @@ members = ["crates/*"]
 [profile.dev]
 opt-level = 1
 # increases incremental compilation speed by about 10% while keeping debug info
-split-debuginfo="unpacked"
+split-debuginfo = "unpacked"
 
 # Enable high optimizations for dependencies (incl. Bevy), but not for our code:
 [profile.dev.package."*"]
@@ -46,7 +46,7 @@ rand = "0.8.5"
 rand_distr = "0.4.3"
 itertools = "0.13.0"
 futures-timer = "3.0.3"
-lightyear = { version = "0.18.0", features = ["leafwing"]}
+lightyear = { version = "0.18.0", features = ["leafwing"] }
 serde_with = "3.11.0"
 bevy_hanabi = "0.14.0"
 bitflags = "2.6.0"
@@ -97,9 +97,9 @@ serialization = { path = "crates/serialization" }
 
 [workspace.metadata.dylint]
 libraries = [
-    { git = "https://github.com/jim-works/bevy_lints", branch="master" },
+    # { git = "https://github.com/jim-works/bevy_lints", branch = "master" },
     # or, if you have the library cloned locally
-    # { path = "../bevy_devtools" },
+    # { path = "../bevy_lints" },
 ]
 
 # disable unknown lints warning
