+++ crates/ui_game_over/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ui_game_over"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bevy = { workspace = true }
+
+interfaces = { path = "../interfaces" }
+world = { path = "../world" }
+ui_core = { path = "../ui_core" }
