+++ crates/ui_menu/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ui_menu"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bevy = { workspace = true }
+bevy_simple_text_input = { workspace = true }
+
+engine = { path = "../engine" }
+util = { path = "../util" }
+interfaces = { path = "../interfaces" }
+serialization = { path = "../serialization" }
+ui_core = { path = "../ui_core" }
